优化图书信息管理逻辑；实现图书借阅管理；修复用户信息管理没有对电话号码进行重复性校验的问题

+++ backend/book-manager/src/main/java/org/epsda/bookmanager/pojo/BorrowRecord.java
@@ -22,15 +22,16 @@ import java.time.LocalDateTime;
 @TableName("borrow_record")
 public class BorrowRecord {
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
     @NotNull
-    private Integer userId;
+    private Long userId;
     @NotNull
-    private Integer bookId;
+    private Long bookId;
     private LocalDateTime borrowTime;
     @NotNull
     private LocalDateTime preReturnTime;
     private LocalDateTime realReturnTime;
+    // 状态和罚金不显示在编辑窗口中
     private Integer status;
     private BigDecimal fine;
     private Integer deleteFlag;
